Add tests for calls to non-contract addresses (#231)
* Add test

* Push tests

* Fix tests

* Add more checks

* Refactor tests

+++ ownership-chain/runtime/src/precompiles/mock.rs
@@ -45,7 +45,7 @@ impl frame_system::Config for Runtime {
 	type BlockHashCount = ConstU64<250>;
 	type Version = ();
 	type PalletInfo = PalletInfo;
-	type AccountData = pallet_balances::AccountData<u64>;
+	type AccountData = pallet_balances::AccountData<u128>;
 	type OnNewAccount = ();
 	type OnKilledAccount = ();
 	type SystemWeightInfo = ();
@@ -63,7 +63,7 @@ parameter_types! {
 impl pallet_balances::Config for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	type WeightInfo = ();
-	type Balance = u64;
+	type Balance = u128;
 	type DustRemoval = ();
 	type ExistentialDeposit = ExistentialDeposit;
 	type AccountStore = System;
